-Added journal sections to book

+++ src/main/java/cassunshine/thework/alchemy/elements/ElementPacket.java
@@ -1,8 +1,8 @@
 package cassunshine.thework.alchemy.elements;
 
-public record ElementPacket(Element element, float amount) {
+public record ElementPacket(Element element, int amount) {
     @Override
     public String toString() {
-        return String.format("Element Packet %s %f", element, amount);
+        return String.format("Element Packet %s %d", element, amount);
     }
 }
